Refactoring names, namespaces and line spaces
