+++ TestTicTacToe/GamePlayTests.cs
@@ -1,3 +1,5 @@
+using System;
+using System.IO;
 using TicTacToe;
 using Xunit;
 
@@ -10,7 +12,7 @@ namespace TestTicTacToe
         public void Should_Change_Player_When_SwapPlayer_Is_Called()
         {
             // Arrange 
-            var game = new GameLoop();
+            var game = new GamePlay();
             
             // Act
             var initialPlayer = game.GetPlayer();
@@ -26,7 +28,7 @@ namespace TestTicTacToe
         public void Should_Change_Player_To_The_Initial_Player_When_SwapPlayer_Is_Called_Twice()
         {
             // Arrange 
-            var game = new GameLoop();
+            var game = new GamePlay();
             
             // Act
             var initialPlayer = game.GetPlayer();
@@ -40,8 +42,25 @@ namespace TestTicTacToe
         }
 
         
-        
-        
-        
+        [Fact]
+        public void Should_End_Game_When_PLayer_Inputs_Quit_Command()
+        {
+            // Arrange
+            var stringWriter = new StringWriter();
+            Console.SetOut(stringWriter);
+            
+            var stringReader = new StringReader("q");
+            Console.SetIn(stringReader);
+
+            var game = new GamePlay();
+            
+            // Act
+            game.Play();
+            var actualString = stringWriter.ToString();
+            
+            // Assert
+            Assert.True(actualString.Contains(Constants.GameQuitMessage));
+        }
+
     }
 }
